# swagger/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentation",
    "description": "This is the API documentation for VMess project",
    "version": "1.0.0"
  },
  "host": "localhost:1002",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/searchUser": {
      "get": {
        "tags": [
          "Room"
        ],
        "summary": "Tìm kiếm người dùng",
        "description": "Endpoint để tìm kiếm người dùng bằng email hoặc userName.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/createRoom": {
      "post": {
        "tags": [
          "Room"
        ],
        "summary": "Tạo phòng chat",
        "description": "Endpoint để tạo phòng chat nếu người dùng chưa có phòng chat với người dùng được tìm kiếm.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "targetUserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{roomId}/messages": {
      "get": {
        "tags": [
          "Room"
        ],
        "summary": "Lấy danh sách tin nhắn của phòng",
        "description": "API này lấy danh sách tin nhắn cũ trong phòng chat, sắp xếp từ mới nhất đến cũ nhất, hỗ trợ phân trang.",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/getUserRooms": {
      "get": {
        "tags": [
          "Room"
        ],
        "summary": "Lấy danh sách phòng chat của người dùng",
        "description": "Endpoint để lấy danh sách các phòng chat của người dùng hiện tại.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/editMessage/{messageId}": {
      "put": {
        "tags": [
          "Message"
        ],
        "summary": "Sửa tin nhắn",
        "description": "Endpoint để sửa nội dung tin nhắn trong phòng chat",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newMessageContent": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/deleteMessage/{messageId}": {
      "delete": {
        "tags": [
          "Message"
        ],
        "summary": "Xóa tin nhắn",
        "description": "Endpoint để xóa tin nhắn trong phòng chat",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}